+++ scripts/macros/monsterFeatures/bodak/bodak.js
@@ -1,0 +1,6 @@
+import {auraOfAnnihilation} from './auraOfAnnihilation.js';
+import {deathGaze} from './deathGaze.js';
+export let bodak = {
+    'deathGaze': deathGaze,
+    'auraOfauraOfAnnihilation': auraOfAnnihilation
+}